requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.1
SpeechRecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
streamlit==1.23.1
requests==2.28.2